Detect and collect deprecated fields in the check
This comment provides a function to collect deprecated fields from the
given check, so they can be logged as an error.

Due do the limitations of jsonschema-rs, it is not possible to log
deprecations if the given schema is not valid otherwise.  I wish to
address this via using <https://github.com/santhosh-tekuri/boon> instead
of committing this to the upstream project, since the API of boon is
way better.

Furthermore, it is not possible to detect deprecations in linked types
($ref), due to shortcomings in the used crate.  Currently this
functionality is not needed and might be addressed via moving to boon