Refactor validators to include diagnostic levels
This commit refactors the deprecation and schema validators to use the
diagnostic levels introduced in this commit.  This is done to allow
better feedback to the user.  Furthermore, the Validator trait is moved
to the validation.rs file, but I'd rather move it into the validators
directory, since it belongs to a different domain.

The expression and value validator still need to be refactored